+++ sdk/search/azure-search-documents/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.0.0b2 (Unreleased)
 
+**Breaking Changes**
+- Removed `SearchApiKeyCredential` and now using `AzureKeyCredential` from azure.core.credentials as key credential
 
 ## 1.0.0b1 (2020-03-10)